Move dimod/io namespace under dimod/serialization
This avoids any name conflict with Python's built-in io namespace.
