+++ MANIFEST.in
@@ -1,4 +1,2 @@
-include dimod/io/bqm_json_schema.json
-include dimod/io/sampleset_json_schema.json
-recursive-include dimod/roof_duality/src *.hpp *.cpp
-recursive-include dimod/roof_duality *.cpp *.pyx
+recursive-include dimod/serialization *.json
+recursive-include dimod/roof_duality *.cpp *.pyx *.hpp
