+++ dimod/serialization/bson.py
