+++ dimod/serialization/coo.py
