+++ dimod/serialization/json.py
